MC-1129 Move static methods to helpers.py file
