+++ plugins/orca_security/icon_orca_security/triggers/new_alert/trigger.py
@@ -3,9 +3,10 @@ import time
 from .schema import NewAlertInput, NewAlertOutput, Input, Output, Component
 
 # Custom imports below
-from insightconnect_plugin_runtime.exceptions import PluginException
 from datetime import datetime
 import json
+
+from icon_orca_security.util.helpers import validate_filters
 
 
 class NewAlert(insightconnect_plugin_runtime.Trigger):
@@ -16,7 +17,7 @@ class NewAlert(insightconnect_plugin_runtime.Trigger):
 
     def run(self, params={}):
         interval = params.get(Input.INTERVAL)
-        additional_filters = self.validate_filters(params.get(Input.FILTERS, []))
+        additional_filters = validate_filters(params.get(Input.FILTERS, []))
         filters = [{"field": "state.created_at", "range": {"gte": datetime.now().isoformat()}}]
         filters.extend(additional_filters)
         while True:
@@ -26,25 +27,3 @@ class NewAlert(insightconnect_plugin_runtime.Trigger):
             for alert in alerts:
                 self.send({Output.ALERT: alert})
             time.sleep(interval)
-
-    @staticmethod
-    def validate_filters(filters: list) -> list:
-        new_filters = []
-        for provided_filter in filters:
-            field = provided_filter.get("field")
-            includes = provided_filter.get("includes")
-            excludes = provided_filter.get("excludes")
-            if not field:
-                raise PluginException(
-                    cause=f"The name of the field against which the alerts should be filtered was not specified in the "
-                    f"filter: {provided_filter}.",
-                    assistance="Please provide a field name and try again.",
-                )
-            if not includes and not excludes:
-                raise PluginException(
-                    cause=f"No values were given for the field to include or exclude in the filter: {provided_filter}.",
-                    assistance="Please provide 'includes' or 'excludes' fields and try again.",
-                )
-            if field != "state.created_at":
-                new_filters.append(provided_filter)
-        return new_filters
